Fix PHP Errors for Catalog & Users
Fixed PHP Errors on the catalog items, user profiles & who the user is following. The first two were caused by not being logged in and the last one was caused by the copy of the followers.php file not being edited properly. This issue has been resolved and will be included in the next update.